refactor: extract DI builder functions from lib.rs to state_builders.rs
lib.rs exceeded the 1000-line cap from AGENTS.md. Move all build_*_state()
functions, resolve_extension_agents, and CronServiceTickRef helper into a
dedicated state_builders module. lib.rs: 1029→540, state_builders: 507.
